feat: raytracer/rasterizer parity, web compatibility & cleanup
- Bring the GPU raytracer closer to parity with the rasterizer (true Lambert,
  matching ambient/IBL and DoF parameters, orthographic camera fix, subsurface).
- Add the rt_switcher feature and enable it (and most example assets) on the
  web, including web-compatible skinning.
- Examples, CI and cleanup: chaining in examples, antialiasing demo, remove the
  multi_light/tonemapping-raytracer leftovers, drop the useless MSAA sample
  zero, and assorted warning/CI fixes.

+++ src/builtin/clustered.rs
@@ -198,8 +198,7 @@ impl Clustered {
         }
         // Round up to the next power of two for headroom, clamped to the device's
         // maximum storage-buffer binding size.
-        let max_lights =
-            (ctxt.device.limits().max_storage_buffer_binding_size as u64 / 64).max(1) as u32;
+        let max_lights = (ctxt.device.limits().max_storage_buffer_binding_size / 64).max(1) as u32;
         let cap = needed
             .next_power_of_two()
             .min(max_lights)
